runtime-requirements.txt
influxdb==5.3.1
paho-mqtt==1.6.1
python-dotenv==0.19.2
requests==2.26.0
schedule==1.1.0